requirements.txt
Django==4.1.2
django-ckeditor==6.5.1
django-crispy-forms==1.14.0
crispy-bootstrap5==0.6
django-allauth==0.50.0
